Add in final sim object to taskdoc
Issue: to restart a sim, we need a Simulation object, not a SimulationState (which is what is stored in the RxnCATaskDoc). Given only the latterthere is no way of making the former, so chaining together simulations is not possible. I'm using the 'structure' from the initial_simulation object (which makes sense imo) and passing around the final simulation object all through the full flow of single sim -> parallel sim -> single task doc -> parallel ray -> multi task doc. This doesn't increase the output doc size by much (~0.5MB) but allows to restart a simulation (can consider making this a kwarg of the Maker in the future).